+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Show smaller Imgur images when data saving mode enabled ([#256](https://github.com/Tunous/Dawn/pull/256))
 - Don't show too small reddit previews ([#264](https://github.com/Tunous/Dawn/pull/264))
 - Fix sharing an image multiple times ([#280](https://github.com/Tunous/Dawn/pull/280))
+- Allow grouping of media download notifications ([#285](https://github.com/Tunous/Dawn/pull/285))
+- Workaround invisible download notifications on Android 11 ([#282](https://github.com/Tunous/Dawn/issues/282))
 
 ## [0.9.2] - 2020-06-13